:art: add sender/status/plugins class for bot
