--- .env.example
@@ -1,2 +1,0 @@
-BOT_QQ=
-WEBHOOK_SECRET=el-psy-congroo
