+++ bot/README.md
@@ -1,0 +1,5 @@
+# bot
+
+这只是一个开发用的测试机器人。
+
+如果你需要一个模版，你应该参考 [el-bot-template](https://github.com/ElpsyCN/el-bot-template)。
