+++ package.json
@@ -1,18 +1,20 @@
 {
   "name": "el-bot",
-  "version": "0.2.0",
-  "description": "A quick qq bot for mirai.",
-  "private": true,
+  "version": "0.3.0",
+  "description": "A quick qq bot framework for mirai.",
+  "main": "dist/index.js",
+  "files": [
+    "dist"
+  ],
   "bin": {
-    "el": "dist/src/cli/index.js"
+    "el": "dist/cli/index.js"
   },
-  "main": "index.ts",
-  "types": "index.d.ts",
   "scripts": {
     "build": "npm-run-all --parllel build:*",
     "build:el-bot": "npm run clean && tsc",
     "build:mirai-ts": "cd packages/mirai-ts && tsc",
     "clean": "rm -rf dist",
+    "bot": "cross-env NODE_ENV=dev nodemon --exec ts-node -r tsconfig-paths/register bot/index.ts",
     "dev": "npm run watch:mirai-ts & sleep 2 && cross-env NODE_ENV=dev nodemon --exec ts-node -r tsconfig-paths/register index.ts",
     "install:mirai": "node scripts/install-mirai-ok.js",
     "install:plugins": "git clone https://github.com/ElpsyCN/el-bot-plugins packages/el-bot-plugins",
@@ -40,17 +42,14 @@
     "elpsy",
     "bot",
     "mirai",
-    "qq"
+    "qq",
+    "framework"
   ],
   "author": {
     "name": "云游君",
     "email": "me@yunyoujun.cn",
     "url": "https://www.yunyoujun.cn"
   },
-  "contributors": [
-    "@YunYouJun",
-    "@ADD-SP"
-  ],
   "license": "AGPL-3.0",
   "bugs": {
     "url": "https://github.com/elpsycn/el-bot/issues"
@@ -60,7 +59,6 @@
     "cac": "^6.6.1",
     "chalk": "^4.0.0",
     "dayjs": "^1.8.28",
-    "dotenv": "^8.2.0",
     "feedparser": "^2.2.10",
     "follow-redirects": "^1.11.0",
     "github-webhook-handler": "^1.0.0",
@@ -86,6 +84,7 @@
     "@typescript-eslint/eslint-plugin": "^3.4.0",
     "@typescript-eslint/parser": "^3.4.0",
     "cross-env": "^7.0.2",
+    "dotenv": "^8.2.0",
     "eslint": "^7.3.1",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
