--- plugins/MiraiAPIHTTP/setting.example.yml
@@ -1,20 +1,0 @@
-## 该配置为全局配置，对所有 Session 有效
-
-# 可选，默认值为 0.0.0.0
-host: "0.0.0.0"
-
-# 可选，端口号，默认值为 8080
-port: 4859
-
-# 可选，默认为 el-psy-congroo，建议手动指定，# 不然容易被 NTR
-authKey: el-psy-congroo
-
-# 可选，缓存大小，默认 4096.缓存过小会导致引用回复与撤回消息失败
-cacheSize: 4096
-
-# 可选，是否开启 websocket，默认关闭，建议通过 Session 范围的配置设置
-enableWebsocket: false
-
-# 可选，配置 CORS 跨域，默认为 *，即允许所有域名
-cors:
-  - "*"
