+++ src/index.ts
@@ -1,0 +1,2 @@
+import Bot from "./bot";
+export default Bot;
